Add tests for GPX helper functions
